+++ PlantsvsZombies/Object/Object.py
@@ -6,12 +6,13 @@ Module for an object with properties position, movement, graphocs and width.
 """
 import PlantsvsZombies.Properties as Properties
 
-class Object:
+class myObject:
     def __init__(self):
         self.Position = Properties.Position()
-        self.Movement = Properties.FastMovement()
         self.grfx = "Actor/Zombie/ozzy.jpg"#"rsz_ozzy.jpg"
         self.width = 36
+        self.dying = False
+        self.stillHangingOn = 6
     
     def getPosition(self):
         """
@@ -21,13 +22,12 @@ class Object:
         """
 
         return self.Position.x,self.Position.y
+
+    def isDead(self):
+        self.stillHangingOn -= 1
+        return self.stillHangingOn <= 0
+        
+
     
-    def isHit(self,other):
-        """
-        Checks if the object collides with an Actor
-
-        :param other the Actor
-        :return True if the positions coinside, False otherwise.
-        """
-
-        return other.getPosition()[0] >= self.getPosition()[0] and other.getPosition()[1] == self.getPosition()[1]
+    def isDying(self):
+        return self.dying