+++ PlantsvsZombies/Object/Projectile/Projectile.py
@@ -1,10 +1,20 @@
 from PlantsvsZombies.Object.Object import *
 
-class Projectile(Object):
+class Projectile(myObject):
     def __init__(self):
-        Object.__init__(self)
+        myObject.__init__(self)
         self.Movement = Properties.FastMovement()
+        self.Attack = Properties.Attack()
     
+    def move(self):
+        """
+        Moves the projectile horizontally from left to right
+
+        :return (x,y) new position as a tuple
+        """
+        self.Position.x += self.Movement.move()
+        return self.Position.x,self.Position.y
+
     def isColliding(self,other):
         """
         Checks if the object collides with an Actor
@@ -12,4 +22,11 @@ class Projectile(Object):
         :param other the Actor
         :return True if the positions coinside, False otherwise.
         """
-        return other.getPosition()[0] >= self.getPosition()[0] and other.getPosition()[1] == self.getPosition()[1]
+        return other.getPosition()[0] <= self.getPosition()[0] and other.getPosition()[1] == self.getPosition()[1]
+
+    def attack(self,other):
+        """
+        Attacks target.
+        """
+        damage = self.Attack.attack()
+        other.Health.looseHealth(damage)