+++ PlantsvsZombies/Properties/Attack/ChargedAttack.py
@@ -8,20 +8,20 @@ from PlantsvsZombies.Properties.Attack.Attack import *
 
 class ChargedAttack(Attack):
     def __init__(self):
+        Attack.__init__(self)
         self.chargeTime = 4
         self.charging = 0
+        self.damage = 10
     
-    def attack(self, other):
+    def attack(self):
         '''
-        If attack is charged, target looses health
+        Returns arrack damage. If charging, returns 0.
 
-        :param other the target as an Actor object
         :return True if attack, False if charging
         '''
         if(not self.charging):
-            other.Health.looseHealth(self.damage)
             self.charging = self.chargeTime
-            return True
+            return self.damage
         else:
             self.charging -= 1
-            return False
+            return 0